mmc: sdhci-s3c: fix missing clock for gpio card-detect
Commit 2abeb5c5ded2 (Add clk_(enable/disable) in runtime suspend/resume)
added the capability to stop the clocks when the device is runtime
suspended, but forgot to handle the case of the card-detect using
an external gpio.

Therefore in the case that runtime-pm is enabled, start the io-clock
when a card is inserted and stop it again once it is removed.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>